refactor(event-handler): Move to up to
