+++ docs/index.md
@@ -144,6 +144,7 @@ aws serverlessrepo list-application-versions \
 | [Tracing](./core/tracer) | Decorators and utilities to trace Lambda function handlers, and both synchronous and asynchronous functions
 | [Logger](./core/logger) | Structured logging made easier, and decorator to enrich structured logging with key Lambda context details
 | [Metrics](./core/metrics) | Custom Metrics created asynchronously via CloudWatch Embedded Metric Format (EMF)
+| [Event handler](./core/event_handler) | Event handler decorators for common Lambda events
 | [Middleware factory](./utilities/middleware_factory) | Decorator factory to create your own middleware to run logic before, and after each Lambda invocation
 | [Parameters](./utilities/parameters) | Retrieve parameter values from AWS Systems Manager Parameter Store, AWS Secrets Manager, or Amazon DynamoDB, and cache them for a specific amount of time
 | [Batch processing](./utilities/batch) | Handle partial failures for AWS SQS batch processing
@@ -152,7 +153,6 @@ aws serverlessrepo list-application-versions \
 | [Event source data classes](./utilities/data_classes) | Data classes describing the schema of common Lambda event triggers
 | [Parser](./utilities/parser) | Data parsing and deep validation using Pydantic
 | [Idempotency](./utilities/idempotency) | Idempotent Lambda handler
-| [Event handler](./utilities/event_handler) | Event handler decorators for common Lambda events
 
 ## Environment variables
