+++ tests/functional/event_handler/test_appsync.py
@@ -5,8 +5,8 @@ from pathlib import Path
 
 import pytest
 
+from aws_lambda_powertools.event_handler import AppSyncResolver
 from aws_lambda_powertools.utilities.data_classes import AppSyncResolverEvent
-from aws_lambda_powertools.utilities.event_handler import AppSyncResolver
 from aws_lambda_powertools.utilities.typing import LambdaContext
